requirements.txt
psycopg2-binary
python-dotenv
langchain
langchain-openai
langchain_experimental
sentence_transformers
pypdf